requirement.txt
streamlit
pandas
scikit-learn
joblib
altair
python-dotenv
